Extended repositories methods

+++ backend/internal/domain/product.go
@@ -3,9 +3,9 @@ package domain
 import "time"
 
 type Product struct {
-	ID          int       `db:"id" json:"id"`
-	BrandID     int       `db:"brand_id" json:"brand_id"`
-	CategoryID  int       `db:"category_id" json:"category_id"`
+	ID          int64     `db:"id" json:"id"`
+	BrandID     int64     `db:"brand_id" json:"brand_id"`
+	CategoryID  int64     `db:"category_id" json:"category_id"`
 	Name        string    `db:"name" json:"name"`
 	Description string    `db:"description" json:"description"`
 	Price       float64   `db:"price" json:"price"`
